+++ plugins/modules/power_port.py
@@ -109,7 +109,10 @@ msg:
   type: str
 """
 
-from ansible_collections.networktocode.nautobot.plugins.module_utils.utils import NAUTOBOT_ARG_SPEC
+from ansible_collections.networktocode.nautobot.plugins.module_utils.utils import (
+    NAUTOBOT_ARG_SPEC,
+    TAGS_ARG_SPEC,
+)
 from ansible_collections.networktocode.nautobot.plugins.module_utils.dcim import (
     NautobotDcimModule,
     NB_POWER_PORTS,
@@ -123,6 +126,7 @@ def main():
     Main entry point for module execution
     """
     argument_spec = deepcopy(NAUTOBOT_ARG_SPEC)
+    argument_spec.update(deepcopy(TAGS_ARG_SPEC))
     argument_spec.update(
         dict(
             device=dict(required=True, type="raw"),
@@ -131,7 +135,6 @@ def main():
             allocated_draw=dict(required=False, type="int"),
             maximum_draw=dict(required=False, type="int"),
             description=dict(required=False, type="str"),
-            tags=dict(required=False, type="list", elements="raw"),
         )
     )