+++ plugins/modules/prefix.py
@@ -61,7 +61,9 @@ options:
     version_added: "3.0.0"
   location:
     description:
-      - Location that prefix is associated with
+      - The single location the prefix will be associated to
+      - If you want to associate multiple locations, use the C(prefix_location) module
+      - Using this parameter will override the C(api_version) option to C(2.0)
     required: false
     type: raw
     version_added: "3.0.0"
@@ -210,7 +212,11 @@ msg:
 """
 
 
-from ansible_collections.networktocode.nautobot.plugins.module_utils.utils import NAUTOBOT_ARG_SPEC
+from ansible_collections.networktocode.nautobot.plugins.module_utils.utils import (
+    NAUTOBOT_ARG_SPEC,
+    TAGS_ARG_SPEC,
+    CUSTOM_FIELDS_ARG_SPEC,
+)
 from ansible_collections.networktocode.nautobot.plugins.module_utils.ipam import (
     NautobotIpamModule,
     NB_PREFIXES,
@@ -224,6 +230,8 @@ def main():
     Main entry point for module execution
     """
     argument_spec = deepcopy(NAUTOBOT_ARG_SPEC)
+    argument_spec.update(deepcopy(TAGS_ARG_SPEC))
+    argument_spec.update(deepcopy(CUSTOM_FIELDS_ARG_SPEC))
     argument_spec.update(
         dict(
             ip_version=dict(required=False, type="int"),
@@ -242,8 +250,6 @@ def main():
             ),
             description=dict(required=False, type="str"),
             namespace=dict(required=False, type="str", default="Global"),
-            tags=dict(required=False, type="list", elements="raw"),
-            custom_fields=dict(required=False, type="dict"),
             first_available=dict(required=False, type="bool", default=False),
         )
     )