+++ plugins/modules/virtual_chassis.py
@@ -86,7 +86,10 @@ msg:
   type: str
 """
 
-from ansible_collections.networktocode.nautobot.plugins.module_utils.utils import NAUTOBOT_ARG_SPEC
+from ansible_collections.networktocode.nautobot.plugins.module_utils.utils import (
+    NAUTOBOT_ARG_SPEC,
+    TAGS_ARG_SPEC,
+)
 from ansible_collections.networktocode.nautobot.plugins.module_utils.dcim import (
     NautobotDcimModule,
     NB_VIRTUAL_CHASSIS,
@@ -100,12 +103,12 @@ def main():
     Main entry point for module execution
     """
     argument_spec = deepcopy(NAUTOBOT_ARG_SPEC)
+    argument_spec.update(deepcopy(TAGS_ARG_SPEC))
     argument_spec.update(
         dict(
             name=dict(required=True, type="str"),
             master=dict(required=False, type="raw"),
             domain=dict(required=False, type="str"),
-            tags=dict(required=False, type="list", elements="raw"),
         )
     )