+++ plugins/modules/vlan.py
@@ -26,7 +26,9 @@ extends_documentation_fragment:
 options:
   location:
     description:
-      - The location the VLAN will be associated to
+      - The single location the VLAN will be associated to
+      - If you want to associate multiple locations, use the C(vlan_location) module
+      - Using this parameter will override the C(api_version) option to C(2.0)
     required: false
     type: raw
     version_added: "3.0.0"
@@ -130,7 +132,11 @@ msg:
   type: str
 """
 
-from ansible_collections.networktocode.nautobot.plugins.module_utils.utils import NAUTOBOT_ARG_SPEC
+from ansible_collections.networktocode.nautobot.plugins.module_utils.utils import (
+    NAUTOBOT_ARG_SPEC,
+    TAGS_ARG_SPEC,
+    CUSTOM_FIELDS_ARG_SPEC,
+)
 from ansible_collections.networktocode.nautobot.plugins.module_utils.ipam import (
     NautobotIpamModule,
     NB_VLANS,
@@ -144,6 +150,8 @@ def main():
     Main entry point for module execution
     """
     argument_spec = deepcopy(NAUTOBOT_ARG_SPEC)
+    argument_spec.update(deepcopy(TAGS_ARG_SPEC))
+    argument_spec.update(deepcopy(CUSTOM_FIELDS_ARG_SPEC))
     argument_spec.update(
         dict(
             location=dict(required=False, type="raw"),
@@ -154,8 +162,6 @@ def main():
             status=dict(required=False, type="raw"),
             role=dict(required=False, type="raw"),
             description=dict(required=False, type="str"),
-            tags=dict(required=False, type="list", elements="raw"),
-            custom_fields=dict(required=False, type="dict"),
         )
     )